Hard dependency on Test::CPAN::Changes at runtime in the generated test
If the author choose to use this plugin, he is interested in its result.
So the test must not be skipped if Test::CPAN::Changes is missing
(whatever the reason) as it previously does. So we unconditionnaly load
the module, and a load failure is a failure of the test.
The test is in xt/release/ so we are sure that this change will only
affect the author, not the final user.